Agreeing on how to agree
Solid has grown. While in a small team it was easy to simply talk to each other and come to decisions together, in our larger team this is now can be a little unwieldy. Without creating fatty tissue process for no reason and making sure Solid stays mean and lean, I want to open up a conversation about how to come to decisions together to deliver a cohesive Solid Community Voice. How do we agree to agree? What is the process of making a suggestion and coming to a legitimate conclusion together. How do we overcome differences of opinion to make sure we continue moving forward? 

Let's take what has worked with us, there's no need to reinvent the wheel. Let's write it down so that it is open to anyone who wants to join. Our Solid Community has expanded to include non-developers as well as developers. Non-developers may not be as familiar with the tools we are using so it is important to include resources to quickly explain the necessary parts of the tools we are using to make sure all voices are heard regardless of background.  

On top of the suggesting to the readme on processes, I would like to invite a conversation about label and their descriptions. https://github.com/solid/solid/issues/labels 

Also, who should get which admin privileges and why? https://github.com/orgs/solid/teams What responsibilities do people have per team or admin privilege? 

Lastly, how do we as Solid coordinate with other parties looking to contribute to Solid? For example, pod providers, webid providers, developer tool providers, and Solid app developers.

This is just a suggestion please do feel free to build on it, I am not stuck to this suggestion at all. What matters is that we co-create a process that we can all feel comfortable with that allows our growing community to make decisions with legitimacy.