Create a Gitter Channel Map
On the request from @mediaprophet I've updated the list of gitter channels and their descriptions as I understand them. Please edit if you feel the channel purpose is inaccurate or incomplete. 

If you know what these two channels are for please do say, if there is no purpose then perhaps we should archive them. 
* [community]: 
* [solid-git]:  

Also, the descriptions in the community readme need to be updated in the gitter channel descriptions too.