Rule of Thumb rather than Code of Conduct
People of Solid, 

I have a radical proposal to get rid of the code of conduct completely and replace it with a single rule of thumb ‘Keep to the content.’ 

Why? (you may ask) 

The code of content is stating points that are so obvious that they should not really need to be said. Plus, if we get to a stage where we feel they need to be said, then we have more fundamental social structures that need to be adapted to build a more trusting environment. 

Secondly, if we have rules, somebody has to enforce them, which would (according to the current description of roles) be me as Solid community manager. Realistically speaking, if a person isn’t professional that reflects, more than anything, on their own credibility. 

Blocking people is an unrealistic solution because they can create another profile, so that’s a pretty futile approach.   

What I can do, as community manager, is work hard to contain off-content conversations so that on-topic conversations are more centre stage.

Thoughts?

Mitzi