Update solid-team.md
The pull request to edit the github.com/solid/information suggests to give an overview of the GitHub.com/solid account. The Solid Team is therefore more focused on the Solid specifications because the implementation of the Solid specification is done though projects and the Project scopes and leaders are defined under the GitHub projects.

In practice the user testing is being carried out by the Solid Manager so perhaps should be updated to reflect this. 

The Solid Contributors involved in the implementation of the spec are listed in the W3C wiki as well as the Solid GitHub repositories. The Solid Contributors involved in the specification are listed in the open source licence to recognise their contribution. Solid Contributors have been listed as Solid Experts.