Solid Information
There is a lot of information about Solid in multiple places which means it is difficult for newcomers to navigate. Also multiple wiki lists are emerging in multiple sites meaning the information is diverging and the energy to maintain the wiki being spread thin. 

My proposal would be to have GitHub.com/solid/information as a central place for Solid information. The delicate thing is to get everyone on board who has made other repos covering the same information that could fit into the solid/information repo or at least have links to it. My preference would be to get as much as possible into a single repo to make sure energy doesn't dissipate and spread too thin. The W3C wiki could simply reference GitHub.com/solid/information and focus on the W3C Solid Community Group meeting agenda and planning for the face-to-face. 

The readme is static information and the .md files are lists that need more regular updating. There are several links that would need to be added to .md files once this pull request is merged. 

This is just a proposal, suggestions more than welcome.