Invitation to add and correct
If you see your name please make sure that you agree with the criteria and edit if you feel they are are misrepresentative. 

If you don't see your name and feel you fulfil the criteria in the first row please add your name to a new row on the table and add an x to the columns you feel are relevant to your profile.