Update governance.md
This is important and therefore needs to presented at the W3C Weekly call to give the Solid community the chance to comment as well as the weekly Solid Team meeting for them to incorporate the thoughts of the Solid community. It also needs to be approved through a merge from the Solid Leader for legitimacy.