Update solid-team.md
The 'Solid Team Member' is too vague in its description of what the individuals are responsible for so I tried to detail the specific responsibilities of each of the roles to understand who is accountable for what. Ideally we would have clear descriptions of the responsibilities of each of those individuals independently of each other to that responsibilities don't fall in between two individuals.