Update solid-vision.md
Just to recap the meaning of each of the terms: 
* Vision - A vision statement focuses on tomorrow and what an organization wants to ultimately become. 
* Mission - A mission statement focuses on today and what an organization does to achieve it. 
* Values - The operating philosophies or principles that guide an organisation's internal conduct and relationships

It would be good to make sure we have a common understanding and agreement around these terms. 

In your understanding and own words what do you understand the Solid mission, vision, and values, to be? (no wrong or right answer)?