Freedom, Diversity, and Self-Determination
Self Determination is clearly an important shared value in the Solid community i.e. the idea that the individual should be able to decide what they want and not be squeezed into making a decision by unnecessary bundling of considerations such as keeping  contact details and a wish for privacy. 

Freedom is a very interesting word to replace interoperability because it captures the emotional essence of the drive. Freedom is a loaded word associated very much with US identity ([freedom fries](https://en.wikipedia.org/wiki/Freedom_fries)). I'm not sure it captures the value of Solid which is based on making agreements in the form of standards which could be seen as a form of constraint.  It would be helpful to talk about who has freedom from what because the software providers will have less freedom when they have to comply to standards that allow users the freedom to switch between services. In the rebooting the web of trust 2018 conference there were some conversations around the data that sits in-between two people. So, for example, genetic data. Your decisions around your personal data can influence others. In the case of genetic data, your family. So one person having the freedom to decide about their own data may be in conflict with anther persons freedom. How do we capture the idea that people should have the ability to self determine while not infringing on the freedom of others?

I think 'Diversity' in combination with Self-Determination may capture the Solid values more accurately because it delivers on the idea of offering options that the user should be able to pick from.