Fairness rather than Equality
Perhaps it's an idea to refer to "fairness" rather than "equality" because "fairness" is a term regularly used in GDPR and making the direct link will make the overlap in intention more apparent. I also think that fairness captures the idea that some may have different needs from others and perfect equality can be unfair too.