Efficiency
One emotion that I see spoken about frequently in the Solid channels is this idea that rules need to be kept at a minimum. Rules for the sake of rules should be removed and only the bare essential should be in place to deliver on the core mission and values.