Aim of Standard
Rather than referring to Solid directly as a standard or multiple standards (still a debate I hear on the Solid channels) perhaps we should define the aim of the standard  

I'm wondering how we can be clear about technical standard (or standardS?) being involved while still giving room to design other elements of the solution too. 


Section 5 of the latest version of the spec table of contents proposal is around privacy which Ruben mentioned in the W3C Solid Community Group meeting would likely need a legal component. 

 Maybe something that needs clarifying too. I'd be really interested in making sure the explanation of Solid is clear yet gives space to further development.